table changes to support prices as integers

+++ backend/db/models/spot.js
@@ -65,7 +65,7 @@ module.exports = (sequelize, DataTypes) => {
     },
     price: {
       allowNull: false,
-      type: DataTypes.DECIMAL(4,2)
+      type: DataTypes.INTEGER
     },
     previewUrl: {
       allowNull: false,
